requirements.txt
chatexchange
dulwich
pycrypto
websocket-client
